ref(hub): Move tests out of hub into core package
